+++ object_analytics_rviz/CHANGELOG.rst
@@ -5,3 +5,4 @@ Changelog for package object_analytics_rviz
 0.3.0 (2018-07-03)
 ------------------
 * create marker_publisher.cpp, subscribe detection/localization/tracking and publish 3d box to rviz markers
+* create image_publisher.cpp, subscribe detection/localization/tracking and publish 2d box to rviz image